Fixed redirecting issue by moving the authStateChange listener code to common Header component
