+++ src/components/Header.tsx
@@ -1,8 +1,13 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import { useDispatch } from "react-redux";
+import { Unsubscribe, onAuthStateChanged } from "firebase/auth";
 
 import { NETFLIX_LOGO, USER_AVATAR } from "../utils/constants";
 import { userSignOut } from "../screens/logIn/loginHelper";
+import { addUser, removeUser } from "../redux/slices/userSlice";
+import { auth } from "../config/firebase";
+import { ROUTE_NAMES } from "../navigation/Routes";
 
 interface IProps {
   showSignOut?: boolean;
@@ -12,9 +17,36 @@ const Header = (props: IProps) => {
   const { showSignOut } = props;
 
   const navigate = useNavigate();
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    /**
+     * This is an event listener given by Firebase.
+     * This will be called whenever there is an auth change. This will be called when user signIn, signUp or signOut
+     * By this we don't have to handle of user data in different places. It can be handled in a central place like this
+     *
+     * https://firebase.google.com/docs/reference/js/auth.user
+     */
+    const unSubscribe: Unsubscribe = onAuthStateChanged(auth, (user) => {
+      if (user) {
+        // User is signed in
+        const { email, photoURL, uid, displayName } = user;
+        dispatch(addUser({ email, photoURL, uid, displayName }));
+        navigate(ROUTE_NAMES.BROWSE);
+      } else {
+        // User is signed out
+        dispatch(removeUser());
+        navigate(ROUTE_NAMES.HOME);
+      }
+    });
+
+    return () => {
+      unSubscribe();
+    };
+  }, [dispatch, navigate]);
 
   const handleSigOut = () => {
-    userSignOut(navigate);
+    userSignOut();
   };
 
   return (
